OneDrive/Desktop/web app/backend/requirements.txt
fastapi
python-docx
openai
pandas
matplotlib
